+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - Emit an event when OMEMO has initialized
+- Manually subscribe to device list nodes of contacts without working PEP updates (i.e. missing presence subscription)
 
 ### Fixed
 - Use only strings for data form values used in pubsub publish options and node configuration